server/requirements.txt
fastapi
uvicorn[standard]
pydantic
beanie
motor
python-dotenv
bcrypt
PyJWT
stripe
imagekitio
httpx
google-generativeai
